refactor(redux): Clean css and move components
