+++ samples/css/screen.css
@@ -30,7 +30,7 @@ body {
 	width: 25.6em;
 }
 
-.card {
+.card-items {
 	display: flex;
 	flex-direction: column;
 	box-sizing: border-box;
@@ -44,11 +44,11 @@ body {
 	height: 100%;
 }
 
-.card>:first-child {
+.card-items>:first-child {
 	margin-top: 0.30em;
 }
 
-.card>:last-child {
+.card-items>:last-child {
 	margin-bottom: 0.30em;
 }
