--- src/cards/items/card-item-container.css
@@ -1,37 +1,0 @@
-.card {
-    display: flex;
-    flex-direction: column;
-    box-sizing: border-box;
-    /*
-  background-image: url("data:image/png;base64, iVBORw0KGgoAAAANSUhEUgAAADAAAAAwCAYAAABXAvmHAAAABmJLR0QA/wD/AP+gvaeTAAAACXBIWXMAABcSAAAXEgFnn9JSAAAAB3RJTUUH4QUDDBUmJzvYlAAABxxJREFUaN7tmXlslHUQhp+WLRShQAGxIIecHlSIHArlsAiUW4gnChEFJB74lxqIIYaEiIlHTNQES5BbEAWslEsBAamggOAB1IKgKHLI0SK2nO36h88mG0LLtnQ5Er5k02633++bd+admXdmIXpXDJAA1PD3qFwVonRuIpACjAEGAqeAXKDgagdQScOHAo8Asf49FWgMFAGHgLPlGebyulrq7VZATeAHYKHGDgLaA/8CPwFLgO+vFgCVgRFARzmfD0wD1gEnwyLTERgJVAdOCGAKcPxKAugFPKmBhUAWMFUDqwJDgBuA6eZAPPAE0FP6FgKzgEWXE0Ac0ELDb5cWe4HZwM8a3AEYbjIHBTQdWO/vzYBhQFMjsgf4ENgBnIkmgBbAfUCaXs8G1gCLPetu86Cd1Niql+8Cavl+EfCduZEGdAOSBfoV8KVAyhXATUBfedwY+ANYpvFHTOC+JmoisAlYCnwLnDMiffyZ7+fLBVQD6OL9zT37W+/fd6kAYoABeupm6bNEL/5l0g7Xw3WAw1JpE5CnQZWtPAn+3zDPOmylmgYcA5KA3kYwBjiggxbohFIDaAo8p8cD8vQ9YLefDwAeBqr4fq0JmSu9RgH3+owNQLr8T7BH9LZPnAE+Az6VRo2A56VroY6aDGyLBEAMUM3q0dfD8zx8mUCaWg5vkw4HfcAOoKJNa4Qg8jWqqsZMl+OnTeRRQAMT/w/Pydbj3bQj0eq1EphptILFAWgAjJcOh4DNwFwT8hap1N8D9gJfSKlYoA3wkI3sH+AX4HN7wYMmaqIGzrMPhBK5j+dXMjemWt3idca9UjFfebI3ZHDgPAC19eIJ4C2NuBF4HOgB1NdTK31Qnob1ADrr6S1Wk1UCjQVe04j7rFRjLKkrjMh6I95aI+O0p5FOLpINAatZsQCKfOgZjW9hHjQ0rLOB1cCfRmsYcKfADpqQGwxzsp4PABnmyA+W2EdsZncAPwIfA58IKFcaP2NVq2/+VRdYMNzgwEWqUE2NL7JzFmk8gqxr9VgkLY4ahRc1NF4PNtfQSUZmi/rofqCrtFsu0P4CT/CZi6TU27KB0gAIXUGNf9BeMAnYDowVwEG900clGi+/N0iJTgKarLczjdYq4AWgCfCYkQlR5oDP2RFmA2UBUFHv7LX0JQGvy90MG1myhrc2L3YBHyktMGeGWMFG6PVZfv4y0N2SnCSFMiwAFaxQBZHOA0nq+aCH1Df5fpX/WX6WaFdtb6IN995tSuh04O+wc4+aO7kadJvlto4O2Gyk0oAZGl/fqFQGfgP62XPWGfFSUShGsCfCgPTw1VVjF/v3oyWcs1r5kKJB3aXbxAs4M9Uceb8kw2JLIeSKNPYJNcoUYKfJvFKvHY3gnDx1zlIdUqmY5lrRPAqWFwAMf2hcJGw0PFkGWV5QUnJGepUWQEzY66q4YrnGr+sArgO4SgAEynBPXHkUg8Al3l+gIQ/4/ms768UEYoqzww1hJTh4OQGEumaGjayj8reTQ07WBfRLaO2Y5vCT58ybGRaRqAMIGPbOTkwLnZOzlAYpaqPOCsD13tdOpdpKub1RmbwpbL7uZbcPRANAoSHe4GbhHmCwc+tMDdmhuBvpkJME7HczPURgJ530spyLW7sgq6dQ2+hKJfyZlwzgrCFv6QwwQao86UAzVoGWrmFZwBxFWsBXaPU4zMkuERitoo1ViX7oRBZ0AOpY0jrlYgBC1SHXGbge8Kaz7hwl7UbgKY1oB9yqqsyWChcais65E5robHtMKT3FnKlrcvcyQgfCZHlEA02MHol1Hs4BXrHKdFA+t5G/q4EPBDVawxIiFISVjVRG2OIgtNRKAn6XlgsF2cR7Ys8vvecDOOwNTYFxPiRTnn+jRk+TPilq9Z0CbRhhKTxi4q/zfT/PTZZKGTolx5L7lKvHJJdch0sCsF+Opzni9QfaeuA8t3Lr/SzVzcQuSvdNT67Gt1GWN7YybXUHtV0KDpJKdfX8fJdrh0oCEBThXCkyygc97PSU7kJqghUprwyluzrwtPuhOCvTO2F7pGTgWWlVSYqlO5MXRprERUZjvJ5+zKR71TI3w4pRGjlSpBNGadgpz5hiM6zjc1I19LiryKWXWkbXWHEG67UuNqT55sXBCLkf5wYizyjOM39q21ceFcQ+K9McV5Tl0sgKXC6tdaPWVemQIsAqxZTOEC0r6PVNzs9rfXZPo9LW3FhqVLKjJSV2y8Usk6yTSZhfQtOJM/Hm69lcu/lAm2M1abmgLF8xlUXMnXP/k2OCPyMNCooRfXuAN9RO1YCXbHxVpNO7UvT05ZbTZ/n/u67Nap3eF6BOKHkL9PhQQZ1zjzTDZL5i80BIdM207BaGdfQ4G2K8m7tmVpZs14455TFJRWs9Ms7EDF+H/yzIFeX5oGgBqG7JHWD1ybSDH7vWZu5adtSoLcL+Ax4lB6cKlIwZAAAAAElFTkSuQmCC");
-  /**/
-    background-color: rgba(192, 192, 192, 1);
-    border-radius: 0.30em;
-    padding: 0.15em;
-    width: 100%;
-    height: 100%;
-}
-
-.card>:first-child {
-    margin-top: 0.30em;
-}
-
-.card>:last-child {
-    margin-bottom: 0.30em;
-}
-
-.card-item {
-    background-color: rgba(230, 230, 230, 0.9);
-    border-style: solid;
-    border-radius: 0.30em;
-    border-width: 0.15em;
-    overflow: hidden;
-    margin-top: -0.15em;
-    margin-bottom: 0;
-    margin-left: 0.30em;
-    margin-right: 0.30em;
-    padding-top: 0.15em;
-    padding-bottom: 0.15em;
-    padding-left: 0.30em;
-    padding-right: 0.30em;
-}
