--- src/cards/items/editor/style.css
@@ -1,3 +1,0 @@
-.card-item-editor {
-    float: right;
-}
