+++ src/cards/items/field/CardField.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import styles from './card-field.css';
+import styles from './styles.css';
 
 export const CardField = ({ item, updateItemField }) => {
     return (
