+++ src/cards/items/styles.css
@@ -1,0 +1,35 @@
+.card-items {
+    display: flex;
+    flex-direction: column;
+
+    box-sizing: border-box;
+    background-color: rgba(192, 192, 192, 1);
+    border-radius: 0.30em;
+    padding: 0.15em;
+    width: 100%;
+    height: 100%;
+}
+
+.card-items > :first-child {
+    margin-top: 0.30em;
+}
+
+.card-items > :last-child {
+    margin-bottom: 0.30em;
+}
+
+.card-item {
+    background-color: rgba(230, 230, 230, 0.9);
+    border-radius: 0.30em;
+    border: 0.15em solid;
+    margin: -0.15em 0.30em 0;
+    padding: 0.15em 0.30em;
+
+    display: flex;
+    flex-direction: column;
+}
+
+.card-item-content {
+    overflow: hidden;
+}
+
