+++ src/cards/items/text/CardText.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import styles from './card-text.css';
+import styles from './styles.css';
 
 export const CardText = ({ item, updateItemField }) => {
     return (
