+++ src/cards/items/toolbar/CardItemToolbar.js
@@ -3,23 +3,22 @@ import PropTypes from 'prop-types';
 
 import CardItemTypeSelector from '../types/CardItemTypeSelector';
 
-import styles from './style.css';
+import styles from './styles.css';
 
-export const CardItemEditor = ({ card, item, updateItemField, removeCardItem }) => {
+export const CardItemToolbar = ({ card, item, updateItemField, removeCardItem }) => {
     return (
-        <div className={styles['card-item-editor']}>
+        <div className={styles['card-item-toolbar']}>
             <CardItemTypeSelector card={card} item={item} updateItemField={updateItemField}/>
             <button type="button" onClick={removeCardItem}>-</button>
         </div>
     );
-
 };
 
-CardItemEditor.propTypes = {
+CardItemToolbar.propTypes = {
     card: PropTypes.object.isRequired,
     item: PropTypes.object.isRequired,
     updateItemField: PropTypes.func.isRequired,
     removeCardItem: PropTypes.func.isRequired
 };
 
-export default CardItemEditor;
+export default CardItemToolbar;
