+++ src/cards/toolbar/CardToolbar.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import styles from './styles.css';
+
 const CardToolbar = ({ duplicate, remove, addCardItem }) => {
     return (
-        <div className="card-container-item-toolbar">
+        <div className={styles['card-toolbar']}>
             <button type="button" onClick={() => duplicate()}>Dupliquer</button>
             <button type='button' onClick={() => remove()}>Supprimer</button>
             <button type='button' onClick={() => addCardItem()}>+</button>
