+++ src/cards/toolbar/styles.css
@@ -1,0 +1,4 @@
+.card-toolbar {
+    position: sticky;
+    margin-top: -2em;
+}
