+++ src/css/colors.css
@@ -8,7 +8,7 @@
 .white.card-shape{
 	background-color: rgba(255, 255, 255, 1);
 } 
- .white .card {
+ .white .card-items {
 	background-color: rgba(255, 255, 255, 1);
 }
 
@@ -21,7 +21,7 @@
 	border-color: rgba(255, 255, 255, 1);
 }
  
-.green .card {
+.green .card-items {
 	background-color: rgba(0, 128, 0, 1);
 }
 
@@ -34,7 +34,7 @@
 	border-color: rgba(102, 230, 102, 1);
 }
 
-.green-light .card {
+.green-light .card-items {
 	background-color: rgba(64, 192, 64, 1);
 	background-color: rgba(64, 128, 64, 1);
 }
@@ -48,7 +48,7 @@
 	border-color: rgba(166, 255, 166, 1);
 }
 
-.green-lizard .card {
+.green-lizard .card-items {
 	background-color: rgba(64, 128, 0, 1);
 }
 
@@ -61,7 +61,7 @@
 	border-color: rgba(166, 230, 102, 1);
 }
 
-.blue .card {
+.blue .card-items {
 	background-color: rgba(0, 0, 128, 1);
 	/*color: rgba(230, 230, 230, 1);*/
 }
@@ -75,7 +75,7 @@
 	border-color: rgba(102, 102, 230, 1);
 }
 
-.blue-light .card {
+.blue-light .card-items {
 	background-color: rgba(64, 64, 192, 1);
 }
 
@@ -88,7 +88,7 @@
 	border-color: rgba(166, 166, 255, 1);
 }
 
-.red .card {
+.red .card-items {
 	background-color: rgba(128, 0, 0, 1);
 	/*color: rgba(230, 230, 230, 1);*/
 }
@@ -102,7 +102,7 @@
 	border-color: rgba(205, 102, 102, 1);
 }
 
-.red-light .card {
+.red-light .card-items {
 	background-color: rgba(192, 64, 64, 1);
 }
 
@@ -115,7 +115,7 @@
 	border-color: rgba(230, 192, 192, 1);
 }
 
-.red-gray .card {
+.red-gray .card-items {
 	background-color: rgba(128, 0, 0, 1);
 }
 
@@ -128,7 +128,7 @@
 	border-color: rgba(205, 166, 166, 1);
 }
 
-.cyan .card {
+.cyan .card-items {
 	background-color: rgba(0, 128, 128, 1);
 }
 
@@ -137,7 +137,7 @@
 	border-color: rgba(166, 230, 230, 1);
 }
 
-.magenta .card {
+.magenta .card-items {
 	background-color: rgba(128, 0, 128, 1);
 }
 
@@ -146,7 +146,7 @@
 	border-color: rgba(205, 166, 205, 1);
 }
 
-.yellow .card {
+.yellow .card-items {
 	background-color: rgba(128, 128, 0, 1);
 }
 
@@ -155,7 +155,7 @@
 	border-color: rgba(230, 230, 166, 1);
 }
 
-.yellow-light .card {
+.yellow-light .card-items {
 	/*background-color: rgba(192, 192, 0, 1);*/
 	background-color: rgba(255, 255, 77, 1);
 }
